abstracted puppeteer functions to helper file
